Move 'cli.py' out of 'utils/' to prevent circular imports
We have two options. Either we do this, or we stop "unwrapping" packages
and start using fully-qualified names in imports. This is the cleaner
option.
